requirements.txt
streamlit
pandas
numpy
scipy
scikit-learn
geopy
matplotlib
seaborn
cartopy
pyyaml